Allow using Reform with Shrine::Attacher directly
Previously we required that folks include Shrine's attachment module to
the model. However, some people don't want to pollute their models with
additional methods and want separation of concerns, so they want to be
able to use Shrine::Attacher directly.

With this change the model attachment is assigned only if attachment
module is included.